Remove navigation from checklist item and summary edit pages
