+++ src/components/Box.tsx
@@ -1,13 +1,8 @@
-import React, { FunctionComponent } from 'react';
+import React, { CSSProperties, FunctionComponent } from 'react';
 import styled, { css } from 'styled-components';
 import { FluidObject } from 'gatsby-image';
 
-import {
-  baseSpacer,
-  breakpoints,
-  doubleSpacer,
-  quadrupleSpacer,
-} from '@styles/size';
+import { baseSpacer, breakpoints, doubleSpacer, quadrupleSpacer } from '@styles/size';
 import { white } from '@styles/color';
 import { baseBorderStyle, z1Shadow, z2Shadow, z3Shadow, z4Shadow } from '@styles/mixins';
 
@@ -24,6 +19,7 @@ type BoxProps = {
     };
   };
   onClick?: () => void;
+  style?: CSSProperties;
 };
 
 const renderShadow = (zindex: number) => {
